pyproject.toml
[project]
name = "debugdungeonbackend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.5",
    "djangorestframework>=3.16.1",
]

[dependency-groups]
dev = [
    "ruff>=0.12.8",
]
